+++ src/test/java/com/lstierneyltd/recipebackend/utils/TestConstants.java
@@ -1,5 +1,7 @@
 package com.lstierneyltd.recipebackend.utils;
 
+import java.math.BigDecimal;
+
 public final class TestConstants {
     public static final int ID = 123;
     public static final String RECIPE_1_NAME = "Spaghetti Bolognaise";
@@ -11,7 +13,7 @@ public final class TestConstants {
     public static final String ABBREVIATION = "abbreviation";
     public static final String DESCRIPTION = "description";
     public static final int ORDERING = 420;
-    public static final int QUANTITY = 69;
+    public static final BigDecimal QUANTITY = BigDecimal.valueOf(69);
     public static final int COOKING_TIME = 90;
 
     public static final int TAG_ID = 12346;
